feat: Add lazy loading for weather icon images
The `WeatherModel` struct now includes a new property `iconURL` which is responsible for storing the URL of the weather icon image. The `iconURL` property is now a mutable variable to allow for lazy loading of the image. Additionally, a new private function `getIconURL` has been added to the `WeatherModel` struct to generate the appropriate URL based on the weather icon code.

This change improves the performance of the WeatherApp by only loading the weather icon image when it is needed, resulting in faster page load times.