feat: Add FavoriteLocationView to display favorite locations
The `FavoriteLocationView` struct has been added to the WeatherApp project. This view will be responsible for displaying the user's favorite locations. Currently, it only displays a placeholder text "Hello, World!" but will be updated to show the actual favorite locations in future commits. This addition is part of the ongoing development to enhance the user experience and provide more functionality in the WeatherApp.