chore: Update WeatherModel and WeatherServiceModel
The `WeatherModel` struct has been updated to include a new property `iconURL` for storing the URL of the weather icon image. The `iconURL` property is now mutable to enable lazy loading of the image. Additionally, a private function `getIconURL` has been added to generate the appropriate URL based on the weather icon code.

In the `WeatherServiceModel`, the API key has been removed from the code. Developers should replace it with their own API key.

These changes improve the performance of the WeatherApp by implementing lazy loading for weather icon images and enhancing the security of the API key.