feat: Add FavoriteLocationViewModel.swift and FavoriteLocationView to project
This commit adds the `FavoriteLocationViewModel.swift` file and the `FavoriteLocationView` struct to the WeatherApp project. The `FavoriteLocationViewModel.swift` file is responsible for handling the logic related to favorite locations in the app, while the `FavoriteLocationView` struct is responsible for displaying the user's favorite locations. These additions are part of the ongoing development to enhance the user experience and provide more functionality in the WeatherApp. Additionally, the commit includes updates to the `WeatherModel` and `WeatherServiceModel` to improve the performance of the app by implementing lazy loading for weather icon images and enhancing the security of the API key.