feat: Add MapKitView.swift and SplashScreenView.swift
This commit removes the unused `wallpaper_unix` imageset file and adds the `MapKitView.swift` and `SplashScreenView.swift` files to the WeatherApp project. The `MapKitView.swift` file is responsible for displaying a map view using the MapKit framework, while the `SplashScreenView.swift` file is responsible for showing a splash screen with an animation before displaying the main content. These additions are part of the ongoing development to enhance the user experience and provide more functionality in the WeatherApp.